# Correctly flush L2 (+performance impact & upcoming optimization fork)