@soundOfDestiny 2025-03-16
@ademeure 
What about this pipeline?
- Construct 30 A tensors and 30 B tensors
- Do L2 flush
- Matmul 30 times with no L2 flush