@ademeure 2025-03-16
BTW I should clarify that the current 'main' on my DeeperGEMM repo hasn't been updated in quite a while, I haven't committed most of the optimizations that lead to the results above yet, I'll try to do it very soon (and clean it up after, very much beta quality but not sure I have time to polish it right now).