@soundOfDestiny 2025-03-16
> > @ademeure What about this pipeline?
> > ```
> > * Construct 30 A tensors and 30 B tensors
> > 
> > * Do L2 flush
> > 
> > * Matmul 30 times with no L2 flush
> > ```
> 
> Interesting idea! That makes sense but it depends what you want to measure. There's (at least) 3 different kinds of things in the L2 that can help across runs:
> 
> 1. Matrix A data & scaling factors
> 2. Matrix B data & scaling factors
> 3. Instructions/constants/etc. - this only really matters for very short kernels (i.e. 1 wave without a very large K).
> 
> You'd prevent (1) and (2) from being in the cache (more so than your real workload even, e.g. A might be in the L2 if they are activations generated by the previous kernel?) while it would allow (3) to potentially stay in the L2 across runs. That would reduce the "prologue" time slightly. I'm not sure how noticeable that effect will be here to be honest.
> 
> Same thing for thermal throttling - it depends what you want to measure and the rest of the workload. If you do the above _without_ any sleeps and a short L2 flush (i.e. not the crazy 8GiB L2 flush I've implemented), you'll end up quite thermally limited for the kernels that get closest to peak TFLOPS (i.e. large m/n/k) on many setups.
> 
> What NSight Compute does (by default) is just flush the L2 between every run and reduce the maximum clock speed so it can remain stable (I think it's set to the base clock instead of boost clock). That's 's a fairly standard approach that's easy to explain/replicate, but there's no reason it _needs_ to work that way, as long as the methodology is well explained.

Indeed.
During inference we usually convert a BF16 tensor to FP8 so A might be in the L2, but perhaps sometimes that might not be the case in the backward pass during training.