@soundOfDestiny 2025-03-16
@ademeure Sorry for my misleading.
> // TODO - I'm going to be slightly too honest and admit it's 2AM and I can't remember why I replaced BLOCK_M with 64
         // but otherwise it crashes, so...

That's because you make two threads to issue 2 TMA stores of 2 math WGs, respectively.