@ademeure 2025-03-18
> @ademeure Sorry for my misleading.
> 
> > // TODO - I'm going to be slightly too honest and admit it's 2AM and I can't remember why I replaced BLOCK_M with 64
> > // but otherwise it crashes, so...
> 
> That's because you make two threads to issue 2 TMA stores of 2 math WGs, respectively.

Yes, thank you! :) Trying to add comments while jet lagged wasn't the best idea (but I hope it was funny at least). Obviously that code is very much work-in-progress and I will polish it and improve it further very soon (probably mostly after I am back from GTC next week).

I pushed it early so you could look at it instead of my much older code if you wanted to, and so that I could show/discuss it with people at GTC. I think after some discussions I now have a better idea of how to optimally use TMA tensor loads for the B Matrix L2 side optimization instead of multiple TMA 1D loads. The numbers I showed in my first post were noticeably faster than what you can get with that current code because it only worked for BLOCK_N=128/64/32 and SHAPE_N a nice multiple of that, while the current code is more generic the overhead is much higher. I think it can be made even faster than my earlier code with TMA tensor loads.

Also I wanted to ask you whether/how much you tried to fuse the BF16->FP8 conversion and scaling factor calculations for the output D when it is going to be used as A right after in the forward pass (instead of doing it in a separate kernel)? I think it is definitely possible, but probably needs even better overlap between output of tile X and processing of tile X+1 than what I have in my current code to significantly improve performance.

(Not sure continuing that discussion on this PR makes sense, happy to discuss elsewhere/on the issues page/directly with you if you'd like)